requirements.txt

# Loupe Project
Django==1.3
django-extensions>=0.4.1
django-imagekit>=0.3.1
django-notification==0.1.5
django-registration==0.7
django-registration-defaults==0.3

# Tastypie
django-tastypie==0.9.9
pyyaml==3.10
lxml==2.3
